pyproject.toml
[project]
name = "huggingface-mcp-course"
version = "0.1.0"
description = ""
authors = [
    {name = "Quentin H. Reul",email = "qreul@expert.ai"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

[project.urls]
Homepage = "https://huggingface.co/learn/mcp-course/"
Repository = "https://github.com/qhreul/huggingface-mcp-course"
"Bug Tracker" = "https://github.com/qhreul/huggingface-mcp-course/issues"

[tool.poetry.dependencies]
fastmcp = "^2.10"
gradio = "^5.35"
huggingface_hub = "^0.32"
python-dotenv = "^1"
smolagents = "^1.19"
textblob = "^0.19"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
